# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ad Library Lead Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadConfidence } from './leadConfidence';
import type { LeadReviewStatus } from './leadReviewStatus';

export interface Lead {
  id: number;
  /** @nullable */
  library_id?: string | null;
  /** @nullable */
  advertiser_name?: string | null;
  /** @nullable */
  final_url?: string | null;
  /** @nullable */
  raw_href?: string | null;
  country: string;
  score: number;
  confidence: LeadConfidence;
  needs_review: boolean;
  review_status: LeadReviewStatus;
  reasons: string[];
  /** @nullable */
  source?: string | null;
  /** @nullable */
  ad_start_date?: number | null;
  session_id: number;
  created_at: Date;
}