# lib/api-zod/src/generated/types/recentActivity.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ad Library Lead Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from './lead';
import type { ScrapeJob } from './scrapeJob';

export interface RecentActivity {
  recent_jobs: ScrapeJob[];
  top_leads: Lead[];
}