# lib/api-zod/src/generated/types/savedSearch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ad Library Lead Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedSearch {
  id: number;
  name: string;
  /** @nullable */
  keyword?: string | null;
  page_ids: string[];
  countries: string[];
  created_at: Date;
}