# lib/api-zod/src/generated/types/scoreBucket.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ad Library Lead Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBucket {
  label: string;
  min: number;
  max: number;
  count: number;
}